// src/app/home/models/logProductImp.ts
export class LogProductImp {
  impSerial?: number= null;
  cntComplete?: number= null;
  cntError?: number= null;
  cntRow?: number= null;

  dataSts?: string= null;
  impFilename?: string= null;
  impSupCode?: string= null;
  impSupSerial?: number= null;
  remark1?: string= null;
  sts01?: string= null;
  sts02?: string= null;
  impDate?: Date= null;

  impDateDesc?: string= null;
  cntFlagDesc?: string= null;

}